requirements.txt
numpy==1.15.4
pandas==0.24.2
scikit-learn==0.20.2
scikits.bootstrap==0.3.3
scipy==1.2.0
sklearn==0.0
